type change to Long ix

+++ backend/src/main/java/com/twittvl/backend/follow/FollowRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface FollowRepository extends JpaRepository<Follow, Integer> {
+public interface FollowRepository extends JpaRepository<Follow, Long> {
     Optional<Follow> findByFollowerIdAndFollowedId(Long followerId, Long followedId);
 
     boolean existsByFollowerIdAndFollowedId(Long followerId, Long followedId);
